+++ src/payne/app/app_version.py
@@ -5,6 +5,7 @@ from pathlib import Path
 import shutil
 
 from payne.app import AppVersionMetadata
+from payne.app import app_version_metadata
 from payne.installer import Installer, InstallSource
 from payne.util.file_system import TemporaryDirectory, safe_create
 
@@ -33,14 +34,17 @@ class AppVersion:
         stem_with_version = f"{original.stem}-{self._version}"
         return original.with_stem(stem_with_version).name
 
-    def _install_scripts(self, source_dir: Path, bin_dir: Path) -> Iterator[Path]:
+    def _install_scripts(self, source_dir: Path, bin_dir: Path) -> Iterator[app_version_metadata.Script]:
         bin_dir.mkdir(parents=True, exist_ok=True)
 
         for source_script in source_dir.iterdir():
             script = bin_dir / self._script_file_name(source_script)
             print(f"Installing script {source_script.name} to {script}")
             shutil.move(source_script, script)
-            yield script
+            yield app_version_metadata.Script(
+                script,
+                source_script.name,
+                app_version_metadata.create_hash(script.read_bytes()))
 
     # Installation #############################################################
 
@@ -53,10 +57,8 @@ class AppVersion:
                 temp_bin_dir = temp_dir / "bin"
                 installer.install(source, root, temp_bin_dir, constraints=constraints_file)
 
-                scripts = self._install_scripts(temp_bin_dir, bin_dir)
-
-                metadata = AppVersionMetadata()
-                metadata.scripts.extend(scripts)
+                scripts = list(self._install_scripts(temp_bin_dir, bin_dir))
+                metadata = AppVersionMetadata(scripts)
                 self.write_metadata(metadata)
 
     def uninstall(self):
@@ -67,7 +69,8 @@ class AppVersion:
 
             for script in metadata.scripts:
                 print(f"Uninstall script {script}")
-                script.unlink(missing_ok=True)
+                # TODO verify the hash
+                script.file.unlink(missing_ok=True)
 
         except BaseException:
             # TODO better
@@ -87,4 +90,4 @@ class AppVersion:
 
     def read_metadata(self) -> AppVersionMetadata:
         data = json.loads(self.metadata_file.read_text())
-        return AppVersionMetadata.parse(data)
+        return AppVersionMetadata.load(data)