+++ doc/changelog.md
@@ -1,5 +1,7 @@
 ...:
   * Metadata file renamed from `payne_app.json` to `payne_app-version.json`
+  * Script metadata extended
+  * Schema version added to metadata
 
 0.2.0:
   * Installing applications from PyPI